// lyfebloc-js/src/types.ts
export interface LyfeblocSdkConfig {
  //use a known network or provide an entirely custom config
  network: Network | LyfeblocNetworkConfig;
  rpcUrl: string;
  //overwrite the subgraph url if you don't want to use the lyfebloc labs maintained version
  customSubgraphUrl?: string;
  subgraphQuery?: GraphQLQuery;
  //optionally overwrite parts of the standard AUTO config
  auto?: Partial<LyfeblocSdkAutoConfig>;
  tenderly?: LyfeblocTenderlyConfig;
}

export interface LyfeblocTenderlyConfig {
  accessKey?: string;
  user?: string;
  project?: string;
  blockNumber?: number;
}

export interface LyfeblocSdkAutoConfig {
  //use a built-in service or provide a custom implementation of a TokenPriceService
  //defaults to coingecko
  tokenPriceService: 'coingecko' | 'subgraph' | TokenPriceService;
  //use a built-in service or provide a custom implementation of a PoolDataService
  //defaults to subgraph
  poolDataService: 'subgraph' | PoolDataService;
  //if a custom PoolDataService is provided, on chain balance fetching needs to be handled externally
  //default to true.
  fetchOnChainBalances: boolean;
}

export interface ContractAddresses {
  Reserve: string;
  multicall: string;
  lyfeblocHelpers: string;
  lidoRelayer?: string;
  relayerV3?: string;
  relayerV4?: string;
  gaugeController?: string;
  feeDistributor?: string;
  veLyfe?: string;
  veLyfeProxy?: string;
  protocolFeePercentagesProvider?: string;
  weightedPoolFactory?: string;
  composableStablePoolFactory?: string;
}

export interface LyfeblocNetworkConfig {
  chainId: Network;
  addresses: {
    contracts: ContractAddresses;
    tokens: {
      wrappedNativeAsset: string;
      lbpRaisingTokens?: string[];
      stETH?: string;
      wstETH?: string;
      bal?: string;
      veLyfe?: string;
      bbaUsd?: string;
    };
  };
  tenderly?: LyfeblocTenderlyConfig;
  urls: {
    subgraph: string;
    gaugesSubgraph?: string;
    blockNumberSubgraph?: string;
  };
  pools: {
    wETHwstETH?: PoolReference;
  };
}
// ...
export type PoolReference = {
  id: string;
  address: string;
};
// ...
export interface GraphQLQuery {
  args: GraphQLArgs;
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  attrs: any;
}